Generate datasets from disjointed Unit Square

generate_disjoint_square generates a data frame with two classes from a unit square with varying overlap, but disjointed along the first feature, that is \(X_0 \sim U(0,1)^{d}\) and \(X_1 \sim U(1,2) \times U(\Delta, 1 + \Delta)^{d-1}\).

generate_disjoint_square(n, m, d, Delta)

Arguments

n

a numeric value. The size of class \(X_0\).

m

a numeric value. The size of class \(X_1\).

d

a numeric value. The dimension of the data.

Delta

a numeric value. The overlap parameter between the two classes.

Value

a data frame with two classes.

Author

Jordan Eckert

Examples

library(plotly)
#> Loading required package: ggplot2
#> 
#> Attaching package: ‘plotly’
#> The following object is masked from ‘package:ggplot2’:
#> 
#>     last_plot
#> The following object is masked from ‘package:stats’:
#> 
#>     filter
#> The following object is masked from ‘package:graphics’:
#> 
#>     layout
n <- 100
m <- 25
d <- 3
Delta <- 0

data <- generate_disjoint_square(n, m, d, Delta)
plot_ly(x = data$X1, y=data$X2, z=data$X3, type="scatter3d", mode="markers", color=data$Class)