Generate datasets from Normal distributions

generate_normal_data generates a data frame with two classes from Normal distributions for each feature with varying overlap, that is \(X_0 \sim N(0,1)^{d}\) and \(X_1 \sim N(\Delta, 1)^{d}\) where \(\Delta\) is a random direction some specified Mahalanobis distance away.

generate_normal_data(n, m, d, dist)

Arguments

n

a numeric value. The size of class \(X_0\).

m

a numeric value. The size of class \(X_1\).

d

a numeric value. The dimension of the data.

dist

a numeric value. The Mahalanobis distance between the two classes.

Value

a data frame with two classes.

Author

Jordan Eckert

Examples

library(plotly)
n <- 100
m <- 25
d <- 3
dist <- 1

data <- generate_normal_data(n, m, d, dist)
plot_ly(x = data$X1, y=data$X2, z=data$X3, type="scatter3d", mode="markers", color=data$Class)