Pure and Proper Class Cover Catch Digraph Resampled Classifier

juggle fits a resampled Pure and Proper Class Cover Catch Digraph (PCCCD) classification model.

juggle(
  x,
  y,
  n_model = 30,
  replace = TRUE,
  min_proportion = 0.5,
  tau = 1,
  verbose = FALSE
)

Arguments

x

feature matrix or dataframe.

y

class factor variable.

n_model

an integer. Number of bagged iterations.

replace

a bool. Should replacement be used in data sampling, if true, uses bootstrap sampling.

min_proportion

a numeric. Minimum proportion of samples to use in each model, used only if replace is FALSE.

tau

a numeric between \((0,1]\). Tau parameter for PCCCD.

verbose

a bool. Should progress be printed to console.

Value

an object of "cccd_classifier" which includes:

i_dominant_list

dominant sample indexes.

x_dominant_list

dominant samples from feature matrix, x

radii_dominant_list

Radiuses of the circle for dominant samples

class_names

class names

k_class

number of classes

proportions

proportions each class covered

tau

tau parameter for P-CCCD

Details

Resampling framework for PCCCD.

Author

Jordan Eckert

Examples

n <- 1000
x1 <- runif(n, 1, 10)
x2 <- runif(n, 1, 10)
x <- cbind(x1, x2)
y <- as.factor(ifelse(3 < x1 & x1 < 7 & 3 < x2 & x2 < 7, "A", "B"))

# testing the performance
i_train <- sample(1:n, round(n*0.8))

x_train <- x[i_train,]
y_train <- y[i_train]

x_test <- x[-i_train,]
y_test <- y[-i_train]

m_pcccd <- pcccd(x = x_train, y = y_train, tau = 1)
pred <- classify_pcccd(pcccd = m_pcccd, newdata = x_test)

juggle_pcccd <- juggle(x = x_train, y = y_train, n_model = 30)
pred2 <- classify_juggle(object = juggle_pcccd, newdata = x_test)

# confusion matrix
table(y_test, pred)
#>       pred
#> y_test   A   B
#>      A  33   1
#>      B   1 165
table(y_test, pred2)
#>       pred2
#> y_test   A   B
#>      A  33   1
#>      B   1 165

# test accuracy
sum(y_test == pred)/nrow(x_test)
#> [1] 0.99
sum(y_test == pred2)/nrow(x_test)
#> [1] 0.99