Pure-CCCD Classifier

Fits a Pure Class Cover Catch Digraph (P-CCCD) classification model. Based on rcccd package function.

mpcccd(x, y, tau = 1)

distance(data, method = 1)

Arguments

x

Either a matrix or data frame containing the training set.

y

A factor containing the class labels for the training set. Note: must be the same length as the number of rows in x.

tau

A real number between \((0,1]\) controlling radii of the ball. Choice does not affect the digraph, but can affect classification performance by constructing bigger balls. Default is 1.

Value

an object of "mpcccd" which includes:

i_dominant_list

dominant sample indexes.

x_dominant_list

dominant samples from feature matrix, x

radii_dominant_list

Radiuses of the circle for dominant samples

cardinality

cardinality of each neighborhood for dominant samples

class_names

class names

k_class

number of classes

Details

Framework for PCCCD. PCCCD determines target class dominant points set \(S\) and their circular cover area by determining balls \(B(x^{\text{target}}, r_{i})\) with radii using minimum amount of dominant point which satisfies \(X^{\text{non-target}}\cap \bigcup_{i} B_{i} = \varnothing\) (pure) and \(X^{\text{target}}\subset \bigcup_{i} B_{i}\) (proper). The calculation of the radii r is based off

\(r(x) := (1- \tau)d(x, l(x)) + \tau d(x, u(x))\) where \(u(x)\) is the argmin distance between x and the non-target points and \(l(x)\) is the argmax distance between x and the other target points such that their distance is less than \(u(x)\).

This guarantees that balls of target class never covers any non-target samples (pure) and balls cover all target samples (proper).

Multiclass classification is handled using a one-versus-rest approach where the non-target class is taken to be the union of all classes that are not the target class.

For further detail, please refer to Priebe et al. (2001), Priebe et al. (2003), and Manukyan and Ceyhan (2016).

Much of the framework is based off the rcccd package with minor changes. Used for Jordan Eckert's dissertation research.

Author

Jordan Eckert